In the GameScreen added method tick
