+++ js/screens/result.js
@@ -3,12 +3,9 @@ import App from '../application';
 import {displayScreen} from '../lib/screenRender';
 
 class ResultScreen {
-  constructor() {
-  }
-
   init(state) {
     this.view = new ResultView(state);
-    window.clearInterval(this.view.state.timer.id);
+    window.clearInterval(this.view.state.timerId);
     this.view.replayHandler = () => {
       App.showWelcome();
     };
